requirements.txt
datasets==2.19.1
google-genai==0.5.0
python-dotenv==1.0.1
sqlalchemy==2.0.31
tabulate==0.9.0
